feat(core): reorganize and rename files
